requirements.txt
streamlit>=1.35.0
requests>=2.31.0
exa-py>=1.1.0
sounddevice>=0.4.6
scipy>=1.11.0
numpy>=1.24.0
python-dotenv>=1.0.1
